SQL-LLM/requirements.txt
streamlit
google-genearativeai
python-dotenv